backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
openai==1.57.4
anthropic==0.40.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
aiofiles==24.1.0
playwright==1.49.1
